refactor(structure): move service/ to application/service/ per review
