+++ backend/application/service/progress_formatter/formatter.go
