+++ backend/application/service/thread_fetcher/fetcher.go
