+++ backend/application/usecase/escalate_question.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 
 	"github.com/Asheze1127/progress-checker/backend/entities"
-	"github.com/Asheze1127/progress-checker/backend/service/slack_notifier"
+	"github.com/Asheze1127/progress-checker/backend/application/service/slack_notifier"
 )
 
 // EscalateQuestionUseCase sets a question status to assigned_mentor and
