+++ backend/application/usecase/login.go
@@ -5,8 +5,8 @@ import (
 	"errors"
 	"fmt"
 
-	"github.com/Asheze1127/progress-checker/backend/service/jwt"
-	"github.com/Asheze1127/progress-checker/backend/service/password_hasher"
+	"github.com/Asheze1127/progress-checker/backend/application/service/jwt"
+	"github.com/Asheze1127/progress-checker/backend/application/service/password_hasher"
 	"github.com/Asheze1127/progress-checker/backend/entities"
 )
