+++ backend/cmd/serve/di.go
@@ -12,15 +12,15 @@ import (
 	"github.com/Asheze1127/progress-checker/backend/api/middleware"
 	"github.com/Asheze1127/progress-checker/backend/api/rest"
 	"github.com/Asheze1127/progress-checker/backend/api/webhook"
-	githubsvc "github.com/Asheze1127/progress-checker/backend/service/github"
-	"github.com/Asheze1127/progress-checker/backend/service/jwt"
-	"github.com/Asheze1127/progress-checker/backend/service/message_queue"
-	"github.com/Asheze1127/progress-checker/backend/service/password_hasher"
-	"github.com/Asheze1127/progress-checker/backend/service/progress_formatter"
-	"github.com/Asheze1127/progress-checker/backend/service/question_sender"
-	"github.com/Asheze1127/progress-checker/backend/service/slack_notifier"
-	"github.com/Asheze1127/progress-checker/backend/service/slack_poster"
-	"github.com/Asheze1127/progress-checker/backend/service/thread_fetcher"
+	githubsvc "github.com/Asheze1127/progress-checker/backend/application/service/github"
+	"github.com/Asheze1127/progress-checker/backend/application/service/jwt"
+	"github.com/Asheze1127/progress-checker/backend/application/service/message_queue"
+	"github.com/Asheze1127/progress-checker/backend/application/service/password_hasher"
+	"github.com/Asheze1127/progress-checker/backend/application/service/progress_formatter"
+	"github.com/Asheze1127/progress-checker/backend/application/service/question_sender"
+	"github.com/Asheze1127/progress-checker/backend/application/service/slack_notifier"
+	"github.com/Asheze1127/progress-checker/backend/application/service/slack_poster"
+	"github.com/Asheze1127/progress-checker/backend/application/service/thread_fetcher"
 	"github.com/Asheze1127/progress-checker/backend/application/usecase"
 	"github.com/Asheze1127/progress-checker/backend/infrastructure/encryption"
 	githubinfra "github.com/Asheze1127/progress-checker/backend/infrastructure/github"
@@ -28,7 +28,7 @@ import (
 	"github.com/Asheze1127/progress-checker/backend/infrastructure/postgres"
 	slackinfra "github.com/Asheze1127/progress-checker/backend/infrastructure/slack"
 	pkgslack "github.com/Asheze1127/progress-checker/backend/pkg/slack"
-	idempotencysvc "github.com/Asheze1127/progress-checker/backend/service/idempotency"
+	idempotencysvc "github.com/Asheze1127/progress-checker/backend/application/service/idempotency"
 	"github.com/Asheze1127/progress-checker/backend/util"
 	"github.com/google/uuid"
 )
